refactor: 구조변경
- TDD 테스트용 클래스 생성
- 패키지 구조 변경 (repository 패키지 추가)
- 사용자 정의 Exception 추가
- 일급컬렉션 Sections 추가 및 로직 변경
- 중복 메서드 제거
