+++ src/main/java/nextstep/subway/line/domain/Sections.java
@@ -1,0 +1,29 @@
+package nextstep.subway.line.domain;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.FetchType;
+import javax.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
+
+@Embeddable
+public class Sections {
+
+    @OneToMany(
+            fetch = FetchType.LAZY,
+            mappedBy = "line",
+            cascade = {CascadeType.PERSIST, CascadeType.MERGE},
+            orphanRemoval = true
+    )
+    private List<Section> sections = new ArrayList<>();
+
+    public void add(Section section) {
+        sections.add(section);
+    }
+
+    public List<Section> getList() {
+        return sections;
+    }
+
+}
