+++ src/main/java/nextstep/subway/line/repository/LineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.repository;
 
+import nextstep.subway.line.domain.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
