+++ src/test/java/nextstep/subway/line/unit/LineUnitTest.java
@@ -1,0 +1,4 @@
+package nextstep.subway.line.unit;
+
+public class LineUnitTest {
+}
